multi-agent/requirements.txt
# Multi-Agent Example dependencies
# The ncp-sdk is the only required dependency
ncp-sdk
